requirements.txt
streamlit
tslearn
python-binance
numpy
pandas
plotly
warnings
pickle
datetime
sklearn
random